+++ CHANGELOG.md
@@ -6,8 +6,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.1.7] - 2018-11-xx
+## [0.1.7] - 2018-11-15
 ### Added
+- `android_database.Cursor#get()` method
+### Changed
+- `android_database.CursorIndexOutOfBoundsException` constructor
 
 ## [0.1.6] - 2018-11-15
 ### Added